% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{add}
\alias{add}
\title{Sum of two vectors}
\usage{
add(x, y, plot = FALSE)
}
\arguments{
\item{x, y}{`Numeric` vectors.}

\item{plot}{`Logical` parameter. If `TRUE` the plot of `y~x` will be shown. Default to `FALSE`}
}
\value{
`Numeric` vector of `x+y`.
}
\description{
This **wonderful** function will calculate for you the _sum of any two numerical vectors_.
}
\details{
The method used to get `x+y` is obvious. No details are needed! The plot also is very  basic.

The user can provide any two vectors `x` and `y` and get `x+y` and the plot of `y~x`.


For more details see this [article from Wikipedia](https://en.wikipedia.org/wiki/Summation).
}
\examples{
# sum of two numbers
add(1, 1)

# sum of two vectors and a plot
add(c(1,2),c(1,8), plot = TRUE)

}
